Add is_atomic flag to control behaviour
- Code cleanups
